+++ src/components/calc/tiers.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import {Price} from "../price/price";
 
 type Props = {
   tiers: Array<any>,
@@ -13,7 +14,7 @@ export function TiersComponent(props: Props){
   const isMaxTier = props.tiers[props.tiers.length - 1].max <= props.count;
 
   return (
-    <ul>
+    <ul className="tier-list">
       {props.tiers.map((item, index) => {
         const isCurrentTier = item.min <= props.count && item.max >= props.count;
         let className = isCurrentTier ? 'text-weight-bold' : '';
@@ -21,7 +22,7 @@ export function TiersComponent(props: Props){
           className = 'text-weight-bold';
         }
 
-        return <li key={item.price} className={className}>{item.min} - {item.max} : {item.price} uah</li>;
+        return <li key={item.price} className={className}>{item.min} - {item.max} : <Price price={item.price}/></li>;
       })}
     </ul>
   );