+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+##### 2020-11-19 0.1.0
+* Add `Price` component to implement the same view of price in all places
+* Change `sizeLabel` and `sizeField` in all form components.
+* Add `size` property for all form components
+
 ##### 2020-11-16 0.0.2
 * Add `CHANGELOG.md` ;)
 * Add `react-router`